app/requirements.txt
python-dotenv
pandas
mecab
git+https://github.com/halcy/Mastodon.py